Removed rails email preview gem to use actionmailer
