+++ Gemfile
@@ -72,7 +72,6 @@ gem 'pg'
 gem 'puma'
 gem 'rack-mini-profiler', require: false
 gem 'rails_autolink'
-gem 'rails_email_preview', '~> 0.2.29'
 gem 'rdiscount'
 gem 'record_tag_helper'
