--- app/mailer_previews/activation_needed_email_preview.rb
@@ -1,8 +1,0 @@
-class ActivationNeededEmailPreview
-  def activation_needed_email
-    user = User.last
-    # arbitrary value for testing purposes
-    user.activation_token = 102
-    UserMailer.activation_needed_email user
-  end
-end
