--- app/mailer_previews/challenge_grade_released_preview.rb
@@ -1,6 +1,0 @@
-class ChallengeGradeReleasedPreview
-  def challenge_grade_released
-    challenge_grade = ChallengeGrade.last
-    NotificationMailer.challenge_grade_released challenge_grade
-  end
-end
