--- app/mailer_previews/earned_badge_awarded_preview.rb
@@ -1,6 +1,0 @@
-class EarnedBadgeAwardedPreview
-  def earned_badge_awarded
-    earned_badge = EarnedBadge.last
-    NotificationMailer.earned_badge_awarded earned_badge
-  end
-end
