--- app/mailer_previews/grade_export_preview.rb
@@ -1,10 +1,0 @@
-class GradeExportPreview
-  def grade_export
-    course = Course.first
-    ExportsMailer.grade_export(
-      course,
-      user = User.first,
-      course.assignments.to_csv
-    )
-  end
-end
