--- app/mailer_previews/grade_released_preview.rb
@@ -1,6 +1,0 @@
-class GradeReleasedPreview
-  def grade_released
-    grade = Grade.last
-    NotificationMailer.grade_released grade.id
-  end
-end
