--- app/mailer_previews/gradebook_export_preview.rb
@@ -1,11 +1,0 @@
-class GradebookExportPreview
-  def gradebook_export
-    course = Course.first
-    ExportsMailer.gradebook_export(
-      course,
-      User.first,
-      "gradebook export", # export type
-      course.assignments.to_csv
-    )
-  end
-end
