--- app/mailer_previews/group_notify_preview.rb
@@ -1,7 +1,0 @@
-class GroupNotifyPreview
-  def group_notify
-    group = Group.first
-    @student = group.students.first
-    NotificationMailer.group_notify group.id
-  end
-end
