--- app/mailer_previews/group_status_updated_preview.rb
@@ -1,7 +1,0 @@
-class GroupStatusUpdatedPreview
-  def group_status_updated
-    group = Group.first
-    @student = group.students.first
-    NotificationMailer.group_status_updated group.id
-  end
-end
