--- app/mailer_previews/lti_error_preview.rb
@@ -1,7 +1,0 @@
-class LTIErrorPreview
-  def lti_error
-    user = User.first
-    course = Course.first
-    NotificationMailer.lti_error user, course
-  end
-end
