--- app/mailer_previews/new_submission_preview.rb
@@ -1,7 +1,0 @@
-class NewSubmissionPreview
-  def new_submission
-    submission = Submission.last
-    professor = User.first
-    NotificationMailer.new_submission submission.id, professor
-  end
-end
