--- app/mailer_previews/revised_submission_preview.rb
@@ -1,7 +1,0 @@
-class RevisedSubmissionPreview
-  def revised_submission
-    submission = Submission.last
-    professor = User.first
-    NotificationMailer.revised_submission submission, professor
-  end
-end
