--- app/mailer_previews/successful_submission_preview.rb
@@ -1,6 +1,0 @@
-class SuccessfulSubmissionPreview
-  def successful_submission
-    submission = Submission.last
-    NotificationMailer.successful_submission submission
-  end
-end
