--- app/mailer_previews/updated_submission_preview.rb
@@ -1,6 +1,0 @@
-class UpdatedSubmissionPreview
-  def updated_submission
-    submission = Submission.last
-    NotificationMailer.updated_submission submission
-  end
-end
