--- app/mailer_previews/welcome_email_preview.rb
@@ -1,6 +1,0 @@
-class WelcomeEmailPreview
-  def welcome_email
-    user = User.last
-    UserMailer.welcome_email user
-  end
-end
