+++ config/environments/development.rb
@@ -7,7 +7,6 @@ Rails.application.configure do
   config.action_mailer.delivery_method = :sendmail
   config.action_mailer.perform_deliveries = true
 
-
   config.action_mailer.raise_delivery_errors = true
   config.action_view.raise_on_missing_translations = true
   config.active_support.deprecation = :log
