+++ test/mailers/previews/announcement_mailer_preview.rb
@@ -1,4 +1,4 @@
-class AnnouncementMailerPreview
+class AnnouncementMailerPreview < ActionMailer::Preview
   def announcement_email
     announcement = @announcement_id ? Announcement.find(@announcement_id) : Announcement.last
     student = @student_id ? User.find(@student_id) : User.with_role_in_course("student", announcement.course).first
